requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
pillow==12.0.0
psycopg2-binary==2.9.11
sqlparse==0.5.3
python-decouple==3.8
django-ckeditor==6.7.0